requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
websockets==12.0

# Data processing and analysis
pandas==2.1.4
numpy==1.26.2
scipy==1.11.4
scikit-learn==1.3.2
statsmodels==0.14.0
arch==6.2.0

# Blockchain and API integrations
web3==6.11.3
requests==2.31.0
aiohttp==3.9.1
python-binance==1.0.19

# Machine Learning and NLP
transformers==4.35.2
torch==2.2.0
nltk==3.8.1
textblob==0.17.1

# Visualization
plotly==5.17.0
dash==2.14.2
dash-bootstrap-components==1.5.0

# Database
sqlalchemy==2.0.23

# Configuration and environment
python-dotenv==1.0.0
pydantic==2.5.0

# Async and concurrency
aiofiles==23.2.1

# Logging and monitoring
loguru==0.7.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1 